fixed keeping of empty directories with .keep files in them

+++ src/main/java/gov/loc/repository/bagit/writer/BagWriter.java
@@ -191,8 +191,9 @@ public class BagWriter {
       for(Entry<Path, String> entry : manifest.getFileToChecksumMap().entrySet()){
         Path relativeLocation = bagRootDir.relativize(entry.getKey());
         Path writeTo = outputDir.resolve(relativeLocation);
-        if(!Files.exists(writeTo)){
-          Files.createDirectories(writeTo.getParent());
+        Path writeToParent = writeTo.getParent();
+        if(!Files.exists(writeTo) && writeToParent != null){
+          Files.createDirectories(writeToParent);
           Files.copy(entry.getKey(), writeTo);
         }
       }
